// frontend/src/api/generated/models/inventory-location-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InventoryLocationItem } from './inventory-location-item';

/**
 * 
 * @export
 * @interface InventoryLocationPageResponse
 */
export interface InventoryLocationPageResponse {
    /**
     * 
     * @type {Array<InventoryLocationItem>}
     * @memberof InventoryLocationPageResponse
     */
    'content': Array<InventoryLocationItem>;
    /**
     * 
     * @type {number}
     * @memberof InventoryLocationPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryLocationPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryLocationPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryLocationPageResponse
     */
    'totalPages': number;
}