// frontend/src/api/generated/models/inventory-product-summary-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StorageCondition } from './storage-condition';

/**
 * 
 * @export
 * @interface InventoryProductSummaryItem
 */
export interface InventoryProductSummaryItem {
    /**
     * 商品ID
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'productId': number;
    /**
     * 商品コード
     * @type {string}
     * @memberof InventoryProductSummaryItem
     */
    'productCode': string;
    /**
     * 商品名
     * @type {string}
     * @memberof InventoryProductSummaryItem
     */
    'productName': string;
    /**
     * 
     * @type {StorageCondition}
     * @memberof InventoryProductSummaryItem
     */
    'storageCondition': StorageCondition;
    /**
     * ケース在庫合計数
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'caseQuantity': number;
    /**
     * ボール在庫合計数
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'ballQuantity': number;
    /**
     * バラ在庫合計数
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'pieceQuantity': number;
    /**
     * 引当数量合計
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'totalAllocatedQty': number;
    /**
     * 有効在庫数合計
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'totalAvailableQty': number;
    /**
     * 総バラ換算数
     * @type {number}
     * @memberof InventoryProductSummaryItem
     */
    'totalPieceEquivalent': number;
}