// frontend/src/api/generated/models/inventory-product-summary-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { InventoryProductSummaryItem } from './inventory-product-summary-item';

/**
 * 
 * @export
 * @interface InventoryProductSummaryPageResponse
 */
export interface InventoryProductSummaryPageResponse {
    /**
     * 
     * @type {Array<InventoryProductSummaryItem>}
     * @memberof InventoryProductSummaryPageResponse
     */
    'content': Array<InventoryProductSummaryItem>;
    /**
     * 
     * @type {number}
     * @memberof InventoryProductSummaryPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryProductSummaryPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryProductSummaryPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof InventoryProductSummaryPageResponse
     */
    'totalPages': number;
}